requirements.txt
requests==2.32.5
pandas==2.3.3
pydantic-settings==2.11.0
python-dotenv==1.1.1
apache-airflow-providers-postgres==6.3.0
apache-airflow-providers-slack==9.3.0
apache-airflow-providers-common-sql==1.28.1








